+++ app/src/main/java/umc/link/zip/domain/model/login/JwtModel.kt
@@ -3,6 +3,6 @@ package umc.link.zip.domain.model.login
 data class JwtModel(
     val userId: Int,
     val nickname: String,
-    val kakaoId: Int,
+    val kakaoId: Long,
     val createdAt: String
 )
